bunch of re-factoring with databse logic

+++ src/main/java/VolunteerLink/Database.java
@@ -15,6 +15,7 @@ import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.MongoCursor;
 import com.mongodb.client.model.Projections;
+import com.mongodb.client.model.Updates;
 import com.mongodb.client.model.Aggregates;
 import static com.mongodb.client.model.Filters.eq;
 
@@ -33,7 +34,7 @@ public class Database {
         try (InputStream input = Database.class.getClassLoader().getResourceAsStream("config.properties")) {
             Properties prop = new Properties();
             if (input == null) {
-                System.out.println("Sorry, unable to find config.properties\nVerify that the config.properties file is in the VolunteerLink/src/main/java/resources folder.");
+                System.out.println("Sorry, unable to find config.properties\nVerify that the config.properties file is in the /src/main/java/resources folder.");
                 return null;
             }
             prop.load(input);
@@ -71,7 +72,7 @@ public class Database {
     }
 
 
-    public MongoDatabase getDatabase() { // This method may need to be removed, all interaction with MongoDB should be done through the "Database" class, though this method is public
+    public MongoDatabase getDatabase() {
         return database;
     }
     // Methods to return collections
@@ -102,6 +103,19 @@ public class Database {
 
         // Delete all id's of the user from assignedUsers fields in eventRolesCollection
         eventRolesCollection.updateMany(eq("assignedUsers", objectId), new Document("$pull", new Document("assignedUsers", objectId)));
+
+        // decrement volunteersRegistered in eventCollection for all events the user is registered for, and increment volunteersNeeded for those events
+        MongoCursor<Document> cursor = eventCollection.find(eq("volunteersRegistered", objectId)).iterator();
+        try {
+            while (cursor.hasNext()) {
+                Document doc = cursor.next();
+                ObjectId eventId = doc.getObjectId("_id");
+                eventCollection.updateOne(eq("_id", eventId), new Document("$inc", new Document("volunteersRegistered", -1)));
+                eventCollection.updateOne(eq("_id", eventId), new Document("$inc", new Document("volunteersNeeded", 1)));
+            }
+        } finally {
+            cursor.close();
+        }
     }
 
     // Work in progress
@@ -165,10 +179,10 @@ public class Database {
         eventCollection.deleteOne(eq("_id", objectId));
 
         // Delete all Event Roles associated with the event
-        eventRolesCollection.deleteMany(eq("eventId", id));
+        eventRolesCollection.deleteMany(eq("eventId", objectId));
 
         // Delete all event role id's from users in their eventRole_id field
-        usersCollection.updateMany(eq("eventRole_id", id), new Document("$unset", new Document("eventRole_id", "")));
+        usersCollection.updateMany(eq("eventRole_id", id), Updates.unset("eventRole_id"));
     }
 
     public void addEventRole(Document eventRole){
@@ -180,53 +194,54 @@ public class Database {
 
         // get the number of assigned users and number needed for the event role to update the event collection
         Document eventRoleDoc = eventRolesCollection.find(eq("_id", objectId)).first();
-        int numberAssigned = eventRoleDoc.getInteger("numberAssigned");
-        int numberNeeded = eventRoleDoc.getInteger("numberNeeded");
-        ObjectId eventId = eventRoleDoc.getObjectId("eventId");
+        if (eventRoleDoc != null) {
+            int numberAssigned = eventRoleDoc.getInteger("numberAssigned");
+            int numberNeeded = eventRoleDoc.getInteger("numberNeeded");
+            ObjectId eventId = eventRoleDoc.getObjectId("eventId");
 
-        eventRolesCollection.deleteOne(eq("_id", objectId));
+            eventRolesCollection.deleteOne(eq("_id", objectId));
 
-        // decrement volunteersRegistered in eventCollection
-        eventCollection.updateOne(eq("_id", eventId), new Document("$inc", new Document("volunteersRegistered", -numberAssigned)));
-        // decrement volunteersNeeded in eventCollection
-        eventCollection.updateOne(eq("_id", eventId), new Document("$inc", new Document("volunteersNeeded", numberNeeded)));
+            // Update event collection to reflect the removal of this event role
+            updateEventCollection(eventId, -numberAssigned, numberNeeded);
 
-        // Delete all id's of the event role from assignedRoles fields in usersCollection
-        usersCollection.updateMany(eq("eventRole_id", objectId), new Document("$unset", new Document("eventRole_id", "")));
-
+            // Delete all id's of the event role from assignedRoles fields in usersCollection
+            usersCollection.updateMany(eq("eventRole_id", id), Updates.unset("eventRole_id"));
+        }
     }
 
-    public void registerForEventRole(ObjectId userId, ObjectId eventRoleId){
-        eventRolesCollection.updateOne(eq("_id", eventRoleId), new Document("$push", new Document("assignedUsers", userId)));
-
-        // increment numberAssigned in eventRolesCollection
-        eventRolesCollection.updateOne(eq("_id", eventRoleId), new Document("$inc", new Document("numberAssigned", 1)));
-        // decrement numberNeeded in eventRolesCollection
-        eventRolesCollection.updateOne(eq("_id", eventRoleId), new Document("$inc", new Document("numberNeeded", -1)));
-
-        // increment volunteersRegistered in eventCollection
-        eventCollection.updateOne(eq("_id", eventRoleId), new Document("$inc", new Document("volunteersRegistered", 1)));
-        // decrement volunteersNeeded in eventCollection
-        eventCollection.updateOne(eq("_id", eventRoleId), new Document("$inc", new Document("volunteersNeeded", -1)));
-
-        // add eventRole_id in usersCollection
-        usersCollection.updateOne(eq("_id", userId), new Document("$set", new Document("eventRole_id", eventRoleId)));
+    public void registerForEventRole(ObjectId userId, ObjectId eventRoleId) {
+        updateEventRoles(eventRoleId, 1, userId);
+        updateEventCollection(eventRoleId, 1, -1);
+        updateUserCollection(userId, eventRoleId, true);
     }
 
-    public void unregisterForEventRole(ObjectId userId, ObjectId eventRoleId){
-        eventRolesCollection.updateOne(eq("_id", eventRoleId), new Document("$pull", new Document("assignedUsers", userId)));
+    public void unregisterForEventRole(ObjectId userId, ObjectId eventRoleId) {
+        updateEventRoles(eventRoleId, -1, userId);
+        updateEventCollection(eventRoleId, -1, 1);
+        updateUserCollection(userId, eventRoleId, false);
+    }
 
-        // decrement numberAssigned in eventRolesCollection
-        eventRolesCollection.updateOne(eq("_id", eventRoleId), new Document("$inc", new Document("numberAssigned", -1)));
-        // increment numberNeeded in eventRolesCollection
-        eventRolesCollection.updateOne(eq("_id", eventRoleId), new Document("$inc", new Document("numberNeeded", 1)));
+    private void updateEventRoles(ObjectId eventRoleId, int delta, ObjectId userId) {
+        eventRolesCollection.updateOne(eq("_id", eventRoleId), Updates.combine(
+            Updates.push("assignedUsers", delta > 0 ? userId : null),
+            Updates.pull("assignedUsers", delta < 0 ? userId : null),
+            Updates.inc("numberAssigned", delta),
+            Updates.inc("numberNeeded", -delta)
+        ));
+    }
 
-        // decrement volunteersRegistered in eventCollection
-        eventCollection.updateOne(eq("_id", eventRoleId), new Document("$inc", new Document("volunteersRegistered", -1)));
-        // increment volunteersNeeded in eventCollection
-        eventCollection.updateOne(eq("_id", eventRoleId), new Document("$inc", new Document("volunteersNeeded", 1)));
+    private void updateEventCollection(ObjectId eventId, int numberAssignedDelta, int numberNeededDelta) {
+        eventCollection.updateOne(eq("_id", eventId), Updates.combine(
+            Updates.inc("volunteersRegistered", numberAssignedDelta),
+            Updates.inc("volunteersNeeded", numberNeededDelta)
+        ));
+    }
 
-        // remove eventRole_id in usersCollection
-        usersCollection.updateOne(eq("_id", userId), new Document("$unset", new Document("eventRole_id", "")));
+    private void updateUserCollection(ObjectId userId, ObjectId eventRoleId, boolean isRegistering) {
+        if (isRegistering) {
+            usersCollection.updateOne(eq("_id", userId), Updates.set("eventRole_id", eventRoleId));
+        } else {
+            usersCollection.updateOne(eq("_id", userId), Updates.unset("eventRole_id"));
+        }
     }
 }
